Refactor interfaces for readability
